requirements.txt
requests==2.28.1
python-dotenv==0.19.2
numpy==1.21.2